+++ include/roq/gateway_status.h
@@ -25,20 +25,19 @@ struct ROQ_PACKED GatewayStatus final {
     DOWNLOADING,
     READY,
     LOGGED_OUT,
-    MAX = LOGGED_OUT
   };
 
-  GatewayStatus() = default;
+  constexpr GatewayStatus() = default;
 
   // cppcheck-suppress noExplicitConstructor
-  inline GatewayStatus(type_t type)  // NOLINT (allow implicit)
+  constexpr GatewayStatus(type_t type)  // NOLINT (allow implicit)
       : type_(type) {}
 
-  inline explicit GatewayStatus(uint8_t type) : type_(validate(type)) {}
+  explicit constexpr GatewayStatus(uint8_t type) : type_(validate(type)) {}
 
-  inline operator type_t() const { return type_; }
+  constexpr operator type_t() const { return type_; }
 
-  inline std::string_view name() const {
+  constexpr std::string_view name() const {
     using namespace roq::literals;
     switch (type_) {
       case type_t::UNDEFINED:
@@ -61,10 +60,10 @@ struct ROQ_PACKED GatewayStatus final {
     return "UNDEFINED"_sv;
   }
 
-  inline operator std::string_view() const { return name(); }
+  constexpr operator std::string_view() const { return name(); }
 
  protected:
-  inline type_t validate(uint8_t type) {
+  constexpr type_t validate(uint8_t type) {
     auto result = static_cast<type_t>(type);
     switch (result) {
       case type_t::UNDEFINED: