+++ include/roq/option_type.h
@@ -17,19 +17,23 @@ namespace roq {
 //! Enumeration of option types
 struct ROQ_PACKED OptionType final {
   //! helper
-  enum type_t : uint8_t { UNDEFINED = 0, CALL, PUT, MAX = PUT };
+  enum type_t : uint8_t {
+    UNDEFINED = 0,
+    CALL,
+    PUT,
+  };
 
-  OptionType() = default;
+  constexpr OptionType() = default;
 
   // cppcheck-suppress noExplicitConstructor
-  inline OptionType(type_t type)  // NOLINT (allow implicit)
+  constexpr OptionType(type_t type)  // NOLINT (allow implicit)
       : type_(type) {}
 
-  inline explicit OptionType(uint8_t type) : type_(validate(type)) {}
+  explicit constexpr OptionType(uint8_t type) : type_(validate(type)) {}
 
-  inline operator type_t() const { return type_; }
+  constexpr operator type_t() const { return type_; }
 
-  inline std::string_view name() const {
+  constexpr std::string_view name() const {
     using namespace roq::literals;
     switch (type_) {
       case type_t::UNDEFINED:
@@ -44,10 +48,10 @@ struct ROQ_PACKED OptionType final {
     return "UNDEFINED"_sv;
   }
 
-  inline operator std::string_view() const { return name(); }
+  constexpr operator std::string_view() const { return name(); }
 
  protected:
-  inline type_t validate(uint8_t type) {
+  constexpr type_t validate(uint8_t type) {
     auto result = static_cast<type_t>(type);
     switch (result) {
       case type_t::UNDEFINED: