+++ include/roq/order_status.h
@@ -26,20 +26,19 @@ struct ROQ_PACKED OrderStatus final {
     WORKING,
     COMPLETED,
     CANCELED,
-    MAX = CANCELED
   };
 
-  OrderStatus() = default;
+  constexpr OrderStatus() = default;
 
   // cppcheck-suppress noExplicitConstructor
-  inline OrderStatus(type_t type)  // NOLINT (allow implicit)
+  constexpr OrderStatus(type_t type)  // NOLINT (allow implicit)
       : type_(type) {}
 
-  inline explicit OrderStatus(uint8_t type) : type_(validate(type)) {}
+  explicit constexpr OrderStatus(uint8_t type) : type_(validate(type)) {}
 
-  inline operator type_t() const { return type_; }
+  constexpr operator type_t() const { return type_; }
 
-  inline std::string_view name() const {
+  constexpr std::string_view name() const {
     using namespace roq::literals;
     switch (type_) {
       case type_t::UNDEFINED:
@@ -64,10 +63,10 @@ struct ROQ_PACKED OrderStatus final {
     return "UNDEFINED"_sv;
   }
 
-  inline operator std::string_view() const { return name(); }
+  constexpr operator std::string_view() const { return name(); }
 
  protected:
-  inline type_t validate(uint8_t type) {
+  constexpr type_t validate(uint8_t type) {
     auto result = static_cast<type_t>(type);
     switch (result) {
       case type_t::UNDEFINED: