+++ include/roq/request_type.h
@@ -22,20 +22,19 @@ struct ROQ_PACKED RequestType final {
     CREATE_ORDER,
     MODIFY_ORDER,
     CANCEL_ORDER,
-    MAX = CANCEL_ORDER
   };
 
-  RequestType() = default;
+  constexpr RequestType() = default;
 
   // cppcheck-suppress noExplicitConstructor
-  inline RequestType(type_t type)  // NOLINT (allow implicit)
+  constexpr RequestType(type_t type)  // NOLINT (allow implicit)
       : type_(type) {}
 
-  inline explicit RequestType(uint8_t type) : type_(validate(type)) {}
+  explicit constexpr RequestType(uint8_t type) : type_(validate(type)) {}
 
-  inline operator type_t() const { return type_; }
+  constexpr operator type_t() const { return type_; }
 
-  inline std::string_view name() const {
+  constexpr std::string_view name() const {
     using namespace roq::literals;
     switch (type_) {
       case type_t::UNDEFINED:
@@ -52,10 +51,10 @@ struct ROQ_PACKED RequestType final {
     return "UNDEFINED"_sv;
   }
 
-  inline operator std::string_view() const { return name(); }
+  constexpr operator std::string_view() const { return name(); }
 
  protected:
-  inline type_t validate(uint8_t type) {
+  constexpr type_t validate(uint8_t type) {
     auto result = static_cast<type_t>(type);
     switch (result) {
       case type_t::UNDEFINED: