+++ include/roq/time_in_force.h
@@ -23,20 +23,19 @@ struct ROQ_PACKED TimeInForce final {
     IOC,  //!< Immediate or Cancel
     GFD,  //!< Good for Day
     GTC,  //!< Good till Canceled
-    MAX = GTC
   };
 
-  TimeInForce() = default;
+  constexpr TimeInForce() = default;
 
   // cppcheck-suppress noExplicitConstructor
-  inline TimeInForce(type_t type)  // NOLINT (allow implicit)
+  constexpr TimeInForce(type_t type)  // NOLINT (allow implicit)
       : type_(type) {}
 
-  inline explicit TimeInForce(uint8_t type) : type_(validate(type)) {}
+  explicit constexpr TimeInForce(uint8_t type) : type_(validate(type)) {}
 
-  inline operator type_t() const { return type_; }
+  constexpr operator type_t() const { return type_; }
 
-  inline std::string_view name() const {
+  constexpr std::string_view name() const {
     using namespace roq::literals;
     switch (type_) {
       case type_t::UNDEFINED:
@@ -55,10 +54,10 @@ struct ROQ_PACKED TimeInForce final {
     return "UNDEFINED"_sv;
   }
 
-  inline operator std::string_view() const { return name(); }
+  constexpr operator std::string_view() const { return name(); }
 
  protected:
-  inline type_t validate(uint8_t type) {
+  constexpr type_t validate(uint8_t type) {
     auto result = static_cast<type_t>(type);
     switch (result) {
       case type_t::UNDEFINED: